requirements.txt
python-telegram-bot==21.6
openai==1.51.0
python-dotenv==1.0.0
aiosqlite==0.20.0
